# Faster reductions over higher dimensions