This provides a near 10x speedup for me on certain tasks, mostly by making things more SIMD-friendly. Demo:
```jl
A = rand(Uint16, 1000, 2000, 50);
maximum(A, 1)
findmax(A, 1)
maximum(A, 3)
findmax(A, 3)
gc()
@time maximum(A, 1)
gc()
@time findmax(A, 1)
gc()
@time maximum(A, 3)
gc()
@time findmax(A, 3)
nothing
```
Master:
```
  0.094944 seconds (164 allocations: 206.184 KB)
  0.092172 seconds (27 allocations: 977.531 KB)
  0.934914 seconds (18 allocations: 3.815 MB)
  1.327757 seconds (33 allocations: 19.075 MB)
```
This PR:
```
  0.094398 seconds (164 allocations: 206.184 KB)
  0.131847 seconds (27 allocations: 977.531 KB)
  0.096865 seconds (18 allocations: 3.815 MB)
  0.211704 seconds (33 allocations: 19.075 MB)
```
This commit works by special-casing the first dimension and hoisting the `min` operation.

As I note in the comments to this commit, I discovered that #9876 doesn't work equally well on all architectures. But these timings are generated on a machine for which the simple demo in #9876 works, so the improvements here seem to be widespread (there's a similar factor of improvement on the other machine). CC @ArchRobison.